refactor: move graphics value types out of the platform port header (T-CORE-GFXTYPES-1)
The package diagram (T-ARCH-GUARD-2) showed the entire style layer depending on
core/platform_api purely to name a Color: Color/TextMetrics/TextStyle (and the
ImageBitmap forward-decl) lived in IGraphicsContext.h, the graphics *port*.

- New dependency-free header core/GraphicsTypes.h holds those value types;
  IGraphicsContext.h includes it and keeps only the interface.
- Repointed the 9 value-only includers (the whole style layer — ComputedStyle.h,
  Stylesheet.h, StyleEngine.cpp, StyleDefaults.cpp, CssParser.cpp — plus
  ColorUtils, TextStyleUtils, TextDecorationUtils, BrowserApp.cpp) to the new
  header. src/style/* no longer depends on core/platform_api at all.
- Fixed one latent transitive-include leak the change exposed
  (TableBorderPainter.cpp used the interface via a cut chain -> direct include).

Regenerated the package + platform_api diagrams: style/* -> platform_api edges
drop from 4 to 0 (style now depends on src/core where the value types live), and
the port class diagram no longer shows Color/TextStyle/TextMetrics. All 496 tests
green; behavior-preserving.
