+++ doc/TODOs.md
@@ -56,7 +56,7 @@ Draft defined in `doc/milestones/milestone9.md` — revalidate scope at kickoff.
 
 ### Tech debt / refactors
 
-- [ ] **[Later P2] T-CORE-GFXTYPES-1: Move Value Types Out Of The Graphics Port Header**; Goal: stop the whole style layer from depending on the platform graphics *port* just to name a `Color`; Scope: `Color`, `TextStyle`, `TextMetrics`, `ImageBitmap` are defined in `src/core/platform_api/IGraphicsContext.h`, so `style/types` (`ComputedStyle`), `style/parser`, and `style/compute` all `#include` the graphics port interface purely for value types (the package diagram shows ~every package depending on `core/platform_api`). Move these POD value types into a dependency-free header (e.g. `core/graphics_types.h` / a `core/types` package); `IGraphicsContext.h` then includes *that* and keeps only the interface. Acceptance: `style/*` no longer depends on `core/platform_api`; the port is pulled only by real graphics drivers (layout for text measurement, renderer, engine, adapters); Tests: existing tests + DependencyFirewall stays green. *(Filed 2026-07-17 from the T-ARCH-GUARD-2 package diagram: the pervasive `-> core/platform_api` edges are mostly value-type coupling, not real port usage.)*
+- [x] **[Later P2] T-CORE-GFXTYPES-1: Move Value Types Out Of The Graphics Port Header** — DONE 2026-07-17. Moved `Color`/`TextMetrics`/`TextStyle` (+ the `ImageBitmap` forward-decl) from `core/platform_api/IGraphicsContext.h` into a new dependency-free `core/GraphicsTypes.h`; the port header now includes that and keeps only the interface. Repointed the 9 value-only includers (the whole style layer — `ComputedStyle.h`, `Stylesheet.h`, `StyleEngine.cpp`, `StyleDefaults.cpp`, `CssParser.cpp` — plus `ColorUtils`, `TextStyleUtils`, `TextDecorationUtils`, `BrowserApp.cpp`) to the new header, so **`src/style/*` no longer depends on `core/platform_api` at all**. One latent transitive-include leak surfaced (`TableBorderPainter.cpp` used the interface via a cut chain) and was fixed with a direct include. All 496 tests green; behavior-preserving.
 - [ ] **[Later P3] T-DOM-STYLE-COUPLING-1: DOM Node Should Not Point At Computed Style**; Goal: remove the inward-depends-outward edge `core/dom -> style/types`; Scope: `DOM::Node` caches a `shared_ptr<ComputedStyle>` for layout/paint convenience, so the innermost layer depends on the style layer (style is conceptually computed *from* the DOM, not vice versa). Consider a side table (node -> computed style) owned by the style/engine layer instead of a member on Node; Acceptance: `core/dom` no longer depends on `style/types`; Tests: style/layout tests unchanged. *(Filed 2026-07-17 from the T-ARCH-GUARD-2 package diagram. Pragmatic shortcut today; lower priority — the change touches every computed-style read site.)*
 
 - [ ] **[Later P2] T-RESOURCE-TYPE-TABLE-1: Data-Drive Resource-Type Behavior**; Goal: make adding a fetched resource type a one-place change instead of shotgun surgery; Scope: today a new `ResourceType` (Document/Stylesheet/Image/Font) threads its behavior mechanically across ~6 layers — the enum (`ResourceStore.h`), request options (`ResourceRequestPlanner`), the loader entry point (`ResourceLoader::request_*`), the update processor's per-type `handle_*` + `ProcessingStats`/`BatchResult` ready-flag, and the store. Introduce a per-type descriptor (e.g. `{binary?, decode-as, mark-ready-on-asset?, ready-flag}`) so the loader/processor dispatch off the table and a new type is mostly one descriptor entry; Acceptance: adding a hypothetical new type (e.g. Script/Media) touches ~1–2 files, not ~6; Tests: existing resource tests stay green; add one asserting the table drives dispatch. *(Filed 2026-07-16 while adding `ResourceType::Font` for T-FONT-FACE-1: the Font path was a clean mechanical mirror of Image at every layer — a sign of consistency, but also that the extension point is boilerplate-heavy. Not rot; pay down at the next natural touch point, e.g. when a 5th resource type lands.)*
