+++ doc/diagrams/hummingbird_packages.puml
@@ -2,13 +2,13 @@
 title Hummingbird - packages (portable layers + platform port)
 package [src] as C_0009392701334512935369 {
 package [core] as C_0014209843741555024820 {
-package [geometry] as C_0003210140419095144248 {
-}
-package [platform_api] as C_0005942173841726501756 {
-}
 package [utils] as C_0013759753959918659013 {
 }
 package [dom] as C_0009904686820845679556 {
+}
+package [geometry] as C_0003210140419095144248 {
+}
+package [platform_api] as C_0005942173841726501756 {
 }
 }
 package [style] as C_0009541372723884481360 {
@@ -72,58 +72,65 @@ package [renderer] as C_0010810530232160936608 {
 package [app] as C_0015197886863879409370 {
 }
 }
+C_0009904686820845679556 ..> C_0003609324848510455853
 C_0005942173841726501756 ..> C_0003210140419095144248
 C_0005942173841726501756 ..> C_0009904686820845679556
 C_0005942173841726501756 ..> C_0016992185117327198002
-C_0013759753959918659013 ..> C_0005942173841726501756
-C_0009904686820845679556 ..> C_0003609324848510455853
-C_0003609324848510455853 ..> C_0005942173841726501756
+C_0003609324848510455853 ..> C_0014209843741555024820
 C_0010750809004078367597 ..> C_0003609324848510455853
-C_0010750809004078367597 ..> C_0005942173841726501756
+C_0010750809004078367597 ..> C_0014209843741555024820
 C_0010750809004078367597 ..> C_0009904686820845679556
 C_0007438010119502517333 ..> C_0003609324848510455853
-C_0007438010119502517333 ..> C_0005942173841726501756
+C_0007438010119502517333 ..> C_0014209843741555024820
 C_0003230105963042407715 ..> C_0010750809004078367597
 C_0003230105963042407715 ..> C_0013759753959918659013
-C_0003230105963042407715 ..> C_0005942173841726501756
+C_0003230105963042407715 ..> C_0014209843741555024820
 C_0008440754822717707571 ..> C_0010750809004078367597
 C_0003262794474333108449 ..> C_0009904686820845679556
 C_0003262794474333108449 ..> C_0010750809004078367597
 C_0016484924776949546325 ..> C_0009904686820845679556
 C_0016484924776949546325 ..> C_0003210140419095144248
-C_0016484924776949546325 ..> C_0005942173841726501756
+C_0016484924776949546325 ..> C_0014209843741555024820
 C_0016484924776949546325 ..> C_0003609324848510455853
-C_0010028313018740234893 ..> C_0005942173841726501756
+C_0016484924776949546325 ..> C_0005942173841726501756
 C_0010028313018740234893 ..> C_0009904686820845679556
 C_0010028313018740234893 ..> C_0003210140419095144248
 C_0010028313018740234893 ..> C_0009643527008220137476
 C_0010028313018740234893 ..> C_0003609324848510455853
+C_0010028313018740234893 ..> C_0000133866933942422160
+C_0010028313018740234893 ..> C_0005942173841726501756
 C_0010028313018740234893 ..> C_0018345349416159714670
-C_0016263449681984960935 ..> C_0005942173841726501756
+C_0016263449681984960935 ..> C_0014209843741555024820
 C_0016263449681984960935 ..> C_0003609324848510455853
+C_0016263449681984960935 ..> C_0005942173841726501756
 C_0016263449681984960935 ..> C_0009643527008220137476
 C_0016263449681984960935 ..> C_0009904686820845679556
 C_0016263449681984960935 ..> C_0003210140419095144248
 C_0016263449681984960935 ..> C_0018345349416159714670
-C_0000133866933942422160 ..> C_0005942173841726501756
+C_0000133866933942422160 ..> C_0010028313018740234893
 C_0000133866933942422160 ..> C_0003210140419095144248
+C_0000133866933942422160 ..> C_0005942173841726501756
 C_0000133866933942422160 ..> C_0003609324848510455853
+C_0000133866933942422160 ..> C_0014209843741555024820
 C_0004106538904384045667 ..> C_0003609324848510455853
 C_0004106538904384045667 ..> C_0001757249062605208465
 C_0004106538904384045667 ..> C_0003210140419095144248
 C_0004106538904384045667 ..> C_0005942173841726501756
 C_0018345349416159714670 ..> C_0003609324848510455853
 C_0018345349416159714670 ..> C_0003210140419095144248
+C_0018345349416159714670 ..> C_0014209843741555024820
 C_0018345349416159714670 ..> C_0005942173841726501756
 C_0018345349416159714670 ..> C_0009904686820845679556
 C_0009643527008220137476 ..> C_0009904686820845679556
-C_0009643527008220137476 ..> C_0005942173841726501756
+C_0009643527008220137476 ..> C_0010028313018740234893
 C_0009643527008220137476 ..> C_0003210140419095144248
 C_0009643527008220137476 ..> C_0000133866933942422160
 C_0009643527008220137476 ..> C_0003609324848510455853
+C_0009643527008220137476 ..> C_0005942173841726501756
 C_0009643527008220137476 ..> C_0018345349416159714670
 C_0001757249062605208465 ..> C_0003210140419095144248
 C_0001757249062605208465 ..> C_0005942173841726501756
+C_0001757249062605208465 ..> C_0014209843741555024820
 C_0001757249062605208465 ..> C_0003609324848510455853
 C_0008363691339124717704 ..> C_0009904686820845679556
 C_0008363691339124717704 ..> C_0005942173841726501756
@@ -149,6 +156,7 @@ C_0016584655481991008975 ..> C_0009643527008220137476
 C_0016584655481991008975 ..> C_0016263449681984960935
 C_0016584655481991008975 ..> C_0003609324848510455853
 C_0017294201864577130835 ..> C_0005942173841726501756
+C_0017294201864577130835 ..> C_0014209843741555024820
 C_0017294201864577130835 ..> C_0009904686820845679556
 C_0017294201864577130835 ..> C_0003210140419095144248
 C_0017294201864577130835 ..> C_0000133866933942422160
@@ -182,11 +190,12 @@ C_0015352416753327187408 ..> C_0005942173841726501756
 C_0015352416753327187408 ..> C_0016484924776949546325
 C_0015352416753327187408 ..> C_0003210140419095144248
 C_0010810530232160936608 ..> C_0003210140419095144248
+C_0010810530232160936608 ..> C_0014209843741555024820
 C_0010810530232160936608 ..> C_0005942173841726501756
 C_0010810530232160936608 ..> C_0016484924776949546325
 C_0015197886863879409370 ..> C_0005942173841726501756
-C_0015197886863879409370 ..> C_0000411124415016547006
 C_0015197886863879409370 ..> C_0014209843741555024820
+C_0015197886863879409370 ..> C_0000411124415016547006
 C_0015197886863879409370 ..> C_0003210140419095144248
 C_0015197886863879409370 ..> C_0005003916588326868713
 C_0015197886863879409370 ..> C_0011270007403136541578
