+++ doc/diagrams/platform_api.puml
@@ -42,34 +42,6 @@ enum C_0012645783935756570189 {
 Curl
 Stub
 }
-class "Color" as C_0001429326908123085830
-class C_0001429326908123085830 {
-__
-+a : unsigned char
-+b : unsigned char
-+g : unsigned char
-+r : unsigned char
-}
-class "TextMetrics" as C_0015039507757522703813
-class C_0015039507757522703813 {
-__
-+ascent : float
-+descent : float
-+height : float
-+underline_position : float
-+underline_thickness : float
-+width : float
-}
-class "TextStyle" as C_0014817139165458707029
-class C_0014817139165458707029 {
-__
-+bold : bool
-+color : Color
-+font_path : std::string
-+font_size : float
-+italic : bool
-+monospace : bool
-}
 abstract "IGraphicsContext" as C_0017612842807371769814
 abstract C_0017612842807371769814 {
 +~IGraphicsContext() constexpr = default : void
@@ -291,11 +263,6 @@ __
 }
 C_0005406508127155539845 o-- C_0004070535993583755214 : +key
 C_0016604046386869755674 o-- C_0011796997026974214122 : +button
-C_0014817139165458707029 o-- C_0001429326908123085830 : +color
-C_0017612842807371769814 ..> C_0001429326908123085830
-C_0017612842807371769814 ..> C_0012817259058340398679
-C_0017612842807371769814 ..> C_0014817139165458707029
-C_0017612842807371769814 ..> C_0015039507757522703813
 C_0012817259058340398679 o-- C_0012455055291288377930 : +format
 C_0001199566768918950911 o-- C_0012817259058340398679 : +frames
 C_0013482843013343585649 ..> C_0012817259058340398679
