+++ src/core/platform_api/IGraphicsContext.h
@@ -2,35 +2,14 @@
 
 #include <cstdint>
 #include <string>
-#include <utility>
 
+#include "core/GraphicsTypes.h"
 #include "core/geometry/Geometry.h"
 
 namespace Hummingbird {
 
-struct ImageBitmap;
-
-struct Color {
-    unsigned char r, g, b, a;
-};
-
-struct TextMetrics {
-    float width = 0.0f;
-    float height = 0.0f;
-    float ascent = 0.0f;
-    float descent = 0.0f;
-    float underline_position = 0.0f;
-    float underline_thickness = 0.0f;
-};
-
-struct TextStyle {
-    std::string font_path;
-    float font_size = 16.0f;
-    bool bold = false;
-    bool italic = false;
-    bool monospace = false;
-    Color color{0, 0, 0, 255};
-};
+// Color / TextMetrics / TextStyle / ImageBitmap now live in core/GraphicsTypes.h
+// so consumers that only need the value types do not depend on this port.
 
 class IGraphicsContext {
 public:
